refactor: 💡 Move code to its own file
Move `generateDictionary()` to its own file to make it testable.
