+++ packages/plugin-testing-integration/src/app.js
@@ -11,9 +11,7 @@ import { JSDOM } from 'jsdom';
 import { requireFromProject, loadFiles } from './helpers';
 import { getConfig } from './configuration';
 import { getBootConfigExtensions } from './bootConfigExtensions';
-import MessageFormat from 'messageformat';
-import globby from 'globby';
-import path from 'path';
+import { generateDictionary } from './localization';
 
 const setIntervalNative = setInterval;
 const setTimeoutNative = setTimeout;
@@ -121,35 +119,6 @@ async function initImaApp(bootConfigMethods = {}) {
     global.$IMA.$App = {};
   }
 
-  function _generateDictionary(languages, locale = 'en') {
-    // FIXME allow setting locale through config
-    const mf = new MessageFormat(locale);
-    const dictionaries = {};
-    const langFileGlobs = languages[locale];
-    globby.sync(langFileGlobs).forEach(file => {
-      try {
-        const filename = path
-          .basename(file)
-          .replace(locale.toUpperCase() + path.extname(file), '');
-
-        const dictJson = requireFromProject(file);
-        const dict = {};
-        Object.keys(dictJson).forEach(key => {
-          const message = dictJson[key];
-          dict[key] = mf.compile(message);
-        });
-
-        dictionaries[filename] = dict;
-      } catch (e) {
-        console.error(
-          `Tried to load dictionary JSON at path "${file}", but recieved following error.`
-        );
-        console.error(e);
-      }
-    });
-    return dictionaries;
-  }
-
   function _installTimerWrappers() {
     global.setInterval = (...args) => {
       let timer = setIntervalNative(...args);
@@ -201,7 +170,7 @@ async function initImaApp(bootConfigMethods = {}) {
 
   vendors = build.vendors;
 
-  global.$IMA.i18n = _generateDictionary(build.languages, config.locale);
+  global.$IMA.i18n = generateDictionary(build.languages, config.locale);
 
   defaultBootConfigMethods = requireFromProject(
     config.appMainPath
